Move staff tags and events functionality to separate cogs
