+++ bot.py
@@ -329,6 +329,8 @@ bot.load_extension("src.discord.ping")
 bot.load_extension("src.discord.staffcommands")
 bot.load_extension("src.discord.staff.invitationals")
 bot.load_extension("src.discord.staff.censor")
+bot.load_extension("src.discord.staff.tags")
+bot.load_extension("src.discord.staff.events")
 bot.load_extension("src.discord.embed")
 bot.load_extension("src.discord.membercommands")
 bot.load_extension("src.discord.devtools")
