+++ src/discord/staffcommands.py
@@ -1031,193 +1031,6 @@ class StaffNonessential(StaffCommands, name="StaffNonesntl"):
         cron_cog = self.bot.get_cog("CronTasks")
         cron_cog.change_bot_status.restart()
 
-    @discord.commands.slash_command(
-        guild_ids = [SLASH_COMMAND_GUILDS],
-        name = "tagadd",
-        description = "Staff command. Adds a new tag."
-    )
-    @permissions.has_any_role(ROLE_STAFF, ROLE_VIP, guild_id = SERVER_ID)
-    async def tag_add(self,
-        ctx,
-        tag_name: Option(str, "The name of the tag to add.", required = True),
-        launch_helpers: Option(str, "Whether launch helpers can use. Defaults to yes.", choices = ["yes", "no"], default = "yes"),
-        members: Option(str, "Whether all members can use this tag. Defaults to yes.", choices = ["yes", "no"], default = "yes")
-        ):
-        commandchecks.is_staff_from_ctx(ctx)
-
-        if tag_name in [t['name'] for t in src.discord.globals.TAGS]:
-            await ctx.interaction.response.send_message(content = f"The `{tag_name}` tag has already been added. To edit this tag, please use `/tagedit` instead.")
-        else:
-            await ctx.interaction.response.defer()
-
-            succesful = False
-            while not succesful:
-                info_message = await ctx.send("Please send the new text for the tag enclosed in a preformatted block. The block should begin and end with three backticks, with no content on the line of the backticks. If no response is found in 2 minutes, the operation will be cancelled.")
-                content_message = await listen_for_response(
-                    follow_id = ctx.user.id,
-                    timeout = 120,
-                )
-
-                if content_message == None:
-                    await ctx.interaction.edit_original_message(content = "No message was found within 2 minutes. Operation cancelled.")
-                    return
-
-                text = content_message.content
-                await content_message.delete()
-                await info_message.delete()
-                matches = re.findall(r"(?<=```\n)(.*)(?=\n```)", text, flags = re.MULTILINE | re.DOTALL)
-                if len(matches) < 0:
-                    await ctx.interaction.edit_original_message(content = "No matching preformatted block was found. Operation cancelled.")
-                    return
-                else:
-                    new_dict = {
-                        'name': tag_name,
-                        'output': matches[0],
-                        'permissions': {
-                            'staff': True,
-                            'launch_helpers': True if launch_helpers == "yes" else False,
-                            'members': True if members == "yes" else False
-                        }
-                    }
-
-                    src.discord.globals.TAGS.append(new_dict)
-                    await insert("data", "tags", new_dict)
-                    succesful = True
-                    await ctx.interaction.edit_original_message(content = f"The `{tag_name}` tag was added!")
-
-    @discord.commands.slash_command(
-        guild_ids = [SLASH_COMMAND_GUILDS],
-        name = "tagedit",
-        description = "Staff command. Edits an existing tag."
-    )
-    @permissions.has_any_role(ROLE_STAFF, ROLE_VIP, guild_id = SERVER_ID)
-    async def tag_edit(self,
-        ctx,
-        tag_name: Option(str, "The tag name to edit the text of.", required = True),
-        launch_helpers: Option(str, "Whether launch helpers can use. Defaults to 'do not change'.", choices = ["yes", "no", "do not change"], default = "do not change"),
-        members: Option(str, "Whether all members can use this tag. Defaults to 'do not change'.", choices = ["yes", "no", "do not change"], default = "do not change")
-        ):
-        commandchecks.is_staff_from_ctx(ctx)
-
-        if tag_name not in [t['name'] for t in src.discord.globals.TAGS]:
-            return await ctx.interaction.response.send_message(content = f"No tag with name `{tag_name}` could be found.")
-
-        tag = [t for t in src.discord.globals.TAGS if t['name'] == tag_name][0]
-
-        await ctx.interaction.response.defer()
-        info_message = await ctx.send(f"The current content of the tag is:\n```\n{tag['output']}\n```\nPlease enclose a new message to associate with the tag by entering a message in a preformatted block (a block of text between three backticks).")
-
-        content_message = await listen_for_response(
-            follow_id = ctx.user.id,
-            timeout = 120,
-        )
-
-        if content_message == None:
-            await ctx.interaction.edit_original_message(content = "No message was found within 2 minutes. Operation cancelled.")
-            return
-
-        text = content_message.content
-        await content_message.delete()
-        await info_message.delete()
-
-        matches = re.findall(r"(?<=```\n)(.*)(?=\n```)", text, flags = re.MULTILINE | re.DOTALL)
-        if len(matches) < 0:
-            await ctx.interaction.edit_original_message(content = "No matching preformatted block was found. Operation cancelled.")
-            return
-        else:
-            update_dict = {}
-
-            tag['output'] = matches[0]
-            update_dict['output'] = matches[0]
-            if launch_helpers != "do not change":
-                tag['permissions']['launch_helpers'] = True if launch_helpers == "yes" else False
-                update_dict['permissions.launch_helpers'] = True if launch_helpers == "yes" else False
-            if members != "do not change":
-                tag['permissions']['members'] = True if members == "yes" else False
-                update_dict['permissions.members'] = True if members == "yes" else False
-
-            await update("data", "tags", tag['_id'], {"$set": update_dict})
-            await ctx.interaction.edit_original_message(content = f"The `{tag_name}` tag was updated.")
-
-    @discord.commands.slash_command(
-        guildids = [SLASH_COMMAND_GUILDS],
-        name = "tagremove",
-        description = "Staff command. Removes a tag completely."
-    )
-    @permissions.has_any_role(ROLE_STAFF, ROLE_VIP, guild_id = SERVER_ID)
-    async def tag_remove(self,
-        ctx,
-        tag_name: Option(str, "The name of the tag to remove.", required = True)
-        ):
-        commandchecks.is_staff_from_ctx(ctx)
-
-        if tag_name not in [t['name'] for t in src.discord.globals.TAGS]:
-            return await ctx.interaction.response.send_message(content = f"No tag with the name of `{tag_name}` was found.")
-
-        tag = [t for t in src.discord.globals.TAGS if t['name'] == tag_name][0]
-        src.discord.globals.TAGS.remove(tag)
-        await delete("data", "tags", tag['_id'])
-        return await ctx.interaction.response.send_message(content = f"The `{tag_name}` tag was deleted.")
-
-    @discord.commands.slash_command(
-        guild_ids = [SLASH_COMMAND_GUILDS],
-        name = "eventadd",
-        description = "Staff command. Adds a new event."
-    )
-    @permissions.has_any_role(ROLE_STAFF, ROLE_VIP, guild_id = SERVER_ID)
-    async def event_add(self,
-        ctx,
-        event_name: Option(str, "The name of the new event.", required = True),
-        event_aliases: Option(str, "The aliases for the new event. Format as 'alias1, alias2'.")
-        ):
-        commandchecks.is_staff_from_ctx(ctx)
-
-        if event_name in [e['name'] for e in src.discord.globals.EVENT_INFO]:
-            return await ctx.interaction.response.send_message(content = f"The `{event_name}` event has already been added.")
-
-        aliases_array = re.findall(r'\w+', event_aliases)
-        new_dict = {
-            'name': event_name,
-            'aliases': aliases_array
-        }
-
-        src.discord.globals.EVENT_INFO.append(new_dict)
-        await insert("data", "events", new_dict)
-
-        server = self.bot.get_guild(SERVER_ID)
-        await server.create_role(
-            name = event_name,
-            color = discord.Color(0x82A3D3),
-            reason = f"Created by {str(ctx.author)} using /eventadd with Pi-Bot."
-            )
-        await ctx.interaction.response.send_message(content = f"The `{event_name}` event was added.")
-
-    @discord.commands.slash_command(
-        guild_ids = [SLASH_COMMAND_GUILDS],
-        name = 'eventremove',
-        description = "Removes an event's availability and optionally, its role from all users."
-    )
-    @permissions.has_any_role(ROLE_STAFF, ROLE_VIP, guild_id = SERVER_ID)
-    async def event_remove(self,
-        ctx,
-        event_name: Option(str, "The name of the event to remove.", required = True),
-        delete_role: Option(str, "Whether to delete the event role from all users. 'no' allows role to remain.", choices = ["no", "yes"], default = "no", required = True)
-        ):
-        commandchecks.is_staff_from_ctx(ctx)
-
-        if event_name not in [e['name'] for e in src.discord.globals.EVENT_INFO]:
-            return await ctx.interaction.response.send_message(content = f"The `{event_name}` event does not exist.")
-
-        if delete_role == "yes":
-            server = self.bot.get_guild(SERVER_ID)
-            role = discord.utils.get(server.roles, name = event_name)
-            await role.delete()
-
-        event = [e for e in src.discord.globals.EVENT_INFO if e['name'] == event_name][0]
-        src.discord.globals.EVENT_INFO.remove(event)
-        await delete("data", "events", event['_id'])
-        await ctx.interaction.response.send_message(content = f"The `{event_name}` event was deleted.")
-
 def setup(bot):
     bot.add_cog(StaffEssential(bot))
     bot.add_cog(StaffNonessential(bot))
